RivuletExpDataProcessing.m changelog: The function for controling output of "findEdges.m" was added. It finds outliers and NaN in each column of the EdgCoords and specifies them in the output for further usage.
findEdges.m changelog:
Main algorithm for lines classification was completely rewritten
and simplified. Now it is capable of better distinction between
edges and can estimate the edge position more accurately based on
modified weighted mean values function.

The function for automatic edge position estimation in case of
more than 1 distinct line was significantly shortened and then
relocated to the main function.

Function plotLines for visualization of the hough transform
results and for manual plate edges set up was slightly modified
to allow user to choose only 1 problematic edge (because if there
is 1 edge with more distinct lines on it, there is no need to
manually specify also the oposite edge)

The program-user communication dialogs were slightly modified
(and cleared, i hope),

rivuletProcessing.m changelog:
Problem with axis in contour visualization from the top was
corrected.

Plotting of the rivulet from the top and of the whole rivulet
profile was joined up into one cycle, because there is no need to
do the same things twice.