Added files for calculation of the maximal local rivulet height and the standard deviation of the profile - in his article from 1995 shetty assumes the shape of the local profile in form of normal distribution defined by the relation:
                          +       2   +
                          |      x    |
 delta(x,z) = delta0(z)exp|- ---------|
                          |          2|
                          +   2*m(z)  +

where delta(x,z) is the local rivulet height on the plate, delta0 is
its maximal height in the local profile and m(z) is the standard
deviation.

This program attempts to find delta0(z), which is the same as
RivHeight obtained from rivuletProcessing.m and m(z) by fitting
the local profiles of the chosen images and to show the evolution
of these two parameters along the plate.

Which is currently working:
- fitting of the parameters, using the Curve Fitting Toolbox
- displaying the parameters and goodnes of the fits in simple graphics

Whis is yet to be done:
- exporting results in the text files
- better (interactive) work with displayed graphics
- improve the fitting - sometimes, the Rsquared obtained are
  inacceptabely low, the initial guess is yet to be imrpoved
  significantly
- add to the program options to change the fitting parameters, such
  as used method etc